Внесены правки: компонент PageHeader венесен как самостоятельый компонент, MenuItems - перенесён в PageHeader
