+++ src/components/PageHeader.vue
@@ -1,9 +1,26 @@
 <template>
-    <div class="header">
+    <div>
+        <div class="header">
         <span>Новокузнечный переулок 4/1</span>
         <span>"8 (812) 123-45-67"</span>
-    </div>
+
+            </div>
+                <MenuItems :items="items"></MenuItems>
+            </div>
 </template>
+
+<script setup>
+import MenuItems from './GUI/MenuItems.vue';
+
+import { ref } from 'vue'
+
+const items = ref([
+  { id: 1, value: 'Каталог' },
+  { id: 2, value: 'Доставка' },
+  { id: 3, value: 'Коллекции' },
+  { id: 4, value: 'Контакты' },
+])
+</script>
 
 <style scoped>
