+++ src/views/MainPage.vue
@@ -1,8 +1,11 @@
 <template>
-<NavigationSection/> 
+<NavigationSection>
+    <PageHeader/>
+</NavigationSection>
 </template>
 
 <script setup>
 import NavigationSection from '../components/NavigationSection.vue';
+import PageHeader from '../components/PageHeader.vue';
 
 </script>
